Move sample plugins to one place
